+++ Rules/IPv4/Outbound/BasicNetworking.ps1
@@ -172,7 +172,7 @@ packet formats and operating semantics as the unicast Domain Name System (DNS)."
 # DHCP (Dynamic Host Configuration Protocol)
 #
 
-New-NetFirewallRule -DisplayName "Dynamic Host Configuration Protocol" `
+New-NetFirewallRule -DisplayName "DHCP Client" `
 	-Platform $Platform -PolicyStore $PolicyStore -Profile $FirewallProfile `
 	-Service Dhcp -Program $ServiceHost -Group $Group `
 	-Enabled True -Action Allow -Direction $Direction -Protocol UDP `
@@ -183,6 +183,36 @@ New-NetFirewallRule -DisplayName "Dynamic Host Configuration Protocol" `
 	-LocalOnlyMapping $false -LooseSourceMapping $false `
 	-Description "Allow DHCPv4 messages for stateful auto-configuration.
 UDP port number 67 is the destination port of a server, and UDP port number 68 is used by the client." `
+	@Logs | Format-Output @Logs
+
+New-NetFirewallRule -DisplayName "DHCP Client (Discovery)" `
+	-Platform $Platform -PolicyStore $PolicyStore -Profile Any `
+	-Service Dhcp -Program $ServiceHost -Group $Group `
+	-Enabled True -Action Allow -Direction $Direction -Protocol UDP `
+	-LocalAddress Any -RemoteAddress 255.255.255.255 `
+	-LocalPort 68 -RemotePort 67 `
+	-LocalUser Any `
+	-InterfaceType $Interface `
+	-LocalOnlyMapping $false -LooseSourceMapping $false `
+	-Description "The DHCP client broadcasts a DHCPDISCOVER message on the network subnet using
+the destination address 255.255.255.255 (limited broadcast) or
+the specific subnet broadcast address (directed broadcast).
+In response to the DHCP offer, the client replies with a DHCPREQUEST message, broadcast to the server,
+requesting the offered address." `
+	@Logs | Format-Output @Logs
+
+New-NetFirewallRule -DisplayName "DHCP Server" `
+	-Platform $Platform -PolicyStore $PolicyStore -Profile Any `
+	-Service Dhcp -Program $ServiceHost -Group $Group `
+	-Enabled False -Action Allow -Direction $Direction -Protocol UDP `
+	-LocalAddress Any -RemoteAddress 255.255.255.255 `
+	-LocalPort 67 -RemotePort 68 `
+	-LocalUser Any `
+	-InterfaceType $Interface `
+	-LocalOnlyMapping $false -LooseSourceMapping $false `
+	-Description "When a DHCP server receives a DHCPDISCOVER message from a client, which is an IP
+address lease request, the DHCP server reserves an IP address for the client and makes a lease offer
+by sending a DHCPOFFER message to the client" `
 	@Logs | Format-Output @Logs
 
 #