+++ Scripts/SetupProfile.ps1
@@ -71,6 +71,9 @@ if (!(Approve-Execute -Accept $Accept -Deny $Deny @Logs)) { exit }
 # Verify permissions to write firewall logs if needed
 & "$ProjectRoot\Scripts\GrantLogs.ps1" $DefaultUser
 
+# NOTE: Reducing the default (4096 KB) log file size makes logs appear quicker but consumes more resources
+$LogSize = 1024
+
 # Setting up profile seem to be slow, tell user what is going on
 Write-Information -Tags "User" -MessageData "INFO: Setting up public firewall profile..." @Logs
 
@@ -78,7 +81,7 @@ Set-NetFirewallProfile -Profile Public -PolicyStore $PolicyStore `
 	-Enabled True -DefaultInboundAction Block -DefaultOutboundAction Block -AllowInboundRules True `
 	-AllowLocalFirewallRules False -AllowLocalIPsecRules False `
 	-NotifyOnListen True -EnableStealthModeForIPsec True -AllowUnicastResponseToMulticast False `
-	-LogAllowed False -LogBlocked True -LogIgnored True -LogMaxSizeKilobytes 1024 `
+	-LogAllowed False -LogBlocked True -LogIgnored True -LogMaxSizeKilobytes $LogSize `
 	-AllowUserApps NotConfigured -AllowUserPorts NotConfigured `
 	-LogFileName "$FirewallLogsFolder\PublicFirewall.log" @Logs
 
@@ -89,7 +92,7 @@ Set-NetFirewallProfile -Profile Private -PolicyStore $PolicyStore `
 	-Enabled True -DefaultInboundAction Block -DefaultOutboundAction Block -AllowInboundRules True `
 	-AllowLocalFirewallRules False -AllowLocalIPsecRules False `
 	-NotifyOnListen True -EnableStealthModeForIPsec True -AllowUnicastResponseToMulticast True `
-	-LogAllowed False -LogBlocked True -LogIgnored True -LogMaxSizeKilobytes 1024 `
+	-LogAllowed False -LogBlocked True -LogIgnored True -LogMaxSizeKilobytes $LogSize `
 	-AllowUserApps NotConfigured -AllowUserPorts NotConfigured `
 	-LogFileName "$FirewallLogsFolder\PrivateFirewall.log" @Logs
 
@@ -100,7 +103,7 @@ Set-NetFirewallProfile -Profile Domain -PolicyStore $PolicyStore `
 	-Enabled True -DefaultInboundAction Block -DefaultOutboundAction Block -AllowInboundRules True `
 	-AllowLocalFirewallRules False -AllowLocalIPsecRules False `
 	-NotifyOnListen True -EnableStealthModeForIPsec True -AllowUnicastResponseToMulticast True `
-	-LogAllowed False -LogBlocked True -LogIgnored True -LogMaxSizeKilobytes 1024 `
+	-LogAllowed False -LogBlocked True -LogIgnored True -LogMaxSizeKilobytes $LogSize `
 	-AllowUserApps NotConfigured -AllowUserPorts NotConfigured `
 	-LogFileName "$FirewallLogsFolder\DomainFirewall.log" @Logs